+++ packages/open-next/src/core/resolve.ts
@@ -21,7 +21,7 @@ export async function resolveConverter<
   if (typeof converter === "function") {
     return converter();
   } else {
-    const m_1 = await import(`../converters/aws-apigw-v2.js`);
+    const m_1 = await import(`../overrides/converters/aws-apigw-v2.js`);
     // @ts-expect-error
     return m_1.default;
   }
