+++ packages/tests-unit/tests/converters/aws-apigw-v1.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable sonarjs/no-duplicate-string */
-import converter from "@opennextjs/aws/converters/aws-apigw-v1.js";
+import converter from "@opennextjs/aws/overrides/converters/aws-apigw-v1.js";
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
 import { Readable } from "stream";
